% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{center_and_standardise}
\alias{center_and_standardise}
\title{Scale data to zero mean and unit variance}
\usage{
center_and_standardise(my_data)
}
\arguments{
\item{my_data}{matrix; rows correspond to records, columns correspond to features.}
}
\value{
matrix; centered and standardised data.
}
\description{
Scale data to zero mean and unit variance
}

