% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{clusters_3d}
\alias{clusters_3d}
\title{3D Clusters}
\usage{
clusters_3d(N, num_cluster = 3)
}
\arguments{
\item{N}{integer; Number of datapoints.}

\item{num_cluster}{integer; number of clusters.}
}
\description{
3D Clusters
}
