% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{corner_planes}
\alias{corner_planes}
\title{Corner Planes}
\usage{
corner_planes(N, angles = 45)
}
\arguments{
\item{N}{integer; Number of datapoints.}

\item{angles}{numeric; Angle between the two planes.}
}
\description{
Corner Planes
}

