% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{gaussian_random_samples}
\alias{gaussian_random_samples}
\title{Randomly sample from Gaussian distribution}
\usage{
gaussian_random_samples(N, sigma = 1)
}
\arguments{
\item{N}{integer; Number of datapoints.}

\item{sigma}{numeric; Controls the variance of the normal distribution.}
}
\description{
Randomly sample from Gaussian distribution
}

