% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Hessian_LLE.R
\name{Hessian_LLE}
\alias{Hessian_LLE}
\title{Hessian Local Linear Embedding}
\usage{
Hessian_LLE(X, k, d)
}
\arguments{
\item{X}{N x D matrix (N samples, D features).}

\item{k}{integer; Number of nearest neighbor.}

\item{d}{integer; The target dimension.}
}
\value{
A list of two objects. The first is the projected data, the second is the mse.
}
\description{
Hessian Local Linear Embedding
}
\details{
Matlab codes were written by David Donoho & Carrie Grimes (2003) and
extracted from Todd Wittman's MANI: Manifold Learning Toolkit.
}
\examples{
#Simulate data
sim_data <- swiss_roll(N = 600)
library(plotly)
p1 <- plotly_3D(sim_data); p1
HLLE_data <- Hessian_LLE(sim_data$data, k = 8, d = 2)
p2 <- plotly_2D(HLLE_data$projection, color = sim_data$colors); p2
}
\references{
Donoho, D. L., & Grimes, C. (2003). Hessian eigenmaps: Locally linear embedding techniques for high-dimensional data. Proceedings of the National Academy of Sciences, 100(10), 5591-5596.

MANI: Manifold Learning Toolkit - http://macs.citadel.edu/wittman/research.html
}

