% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Hessian_LLE.R
\name{HLLE}
\alias{HLLE}
\title{Hessian Local Linear Embedding
Written by David Donoho & Carrie Grimes, 2003.
Extracted from Todd Wittman's MANI: Manifold Learning Toolkit
Source: http://macs.citadel.edu/wittman/research.html
Port from Matlab to R by Jackson Kwok}
\usage{
HLLE(X, k, d)
}
\description{
Hessian Local Linear Embedding
Written by David Donoho & Carrie Grimes, 2003.
Extracted from Todd Wittman's MANI: Manifold Learning Toolkit
Source: http://macs.citadel.edu/wittman/research.html
Port from Matlab to R by Jackson Kwok
}

