% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Laplacian_Eigenmaps.R
\name{Laplacian_Eigenmaps}
\alias{Laplacian_Eigenmaps}
\title{Laplacian Eigenmap}
\usage{
Laplacian_Eigenmaps(X, k, d)
}
\arguments{
\item{X}{N x D matrix (N samples, D features).}

\item{k}{integer; Number of nearest neighbor.}

\item{d}{integer; The target dimension.}
}
\value{
a list of two objects. The first is the projected data
(which are eigenvectors) and the second is the corresponding eigenvalues.
}
\description{
Laplacian Eigenmap
}
\details{
Matlab codes were written by Belkin & Niyogi (2001)
and extracted from Todd Wittman's MANI: Manifold Learning Toolkit.
}
\examples{
#Simulate data
sim_data <- swiss_roll(N = 600)
library(plotly)
p1 <- plotly_3D(sim_data); p1
LE_data <- Laplacian_Eigenmaps(sim_data$data, k = 8, d = 2)
p2 <- plotly_2D(LE_data$eigenvectors, color = sim_data$colors); p2
}
\references{
Belkin, M., & Niyogi, P. (2001, December). Laplacian Eigenmaps and Spectral Techniques for Embedding and Clustering. In NIPS (Vol. 14, pp. 585-591).

MANI: Manifold Learning Toolkit - http://macs.citadel.edu/wittman/research.html
}

