% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laplacian_Eigenmaps.R
\name{leigs}
\alias{leigs}
\title{Laplacian Eigenmap.
Written by Belkin & Niyogi, 2002.
Extracted from Todd Wittman's MANI: Manifold Learning Toolkit
Source: http://macs.citadel.edu/wittman/research.html
Port from Matlab to R by Jackson Kwok}
\usage{
leigs(data, type = "nn", param, ne)
}
\description{
Laplacian Eigenmap.
Written by Belkin & Niyogi, 2002.
Extracted from Todd Wittman's MANI: Manifold Learning Toolkit
Source: http://macs.citadel.edu/wittman/research.html
Port from Matlab to R by Jackson Kwok
}
