% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LLE2.R
\name{LLE2}
\alias{LLE2}
\title{Locally Linear Embedding}
\usage{
LLE2(data, dim = 2, k)
}
\arguments{
\item{data}{N x D matrix (N samples, D features)}

\item{dim}{integer; The target dimension.}

\item{k}{integer; Number of nearest neighbor.}
}
\value{
Projected data; N x dim matrix
}
\description{
Locally Linear Embedding
}
\details{
This is the exact same function as LLE in "RDRToolbox" except that the matrix inverse operation
is replaced by the pseudoinverse operation since the former often has numerical issues.
See "?RDRToolbox::LLE" for more details about usage.
}
\examples{
#Simulate data
sim_data <- swiss_roll(N = 600)
library(plotly)
p1 <- plotly_3D(sim_data); p1
lle_data <- LLE2(sim_data$data, dim = 2, k = 8)
p2 <- plotly_2D(lle_data, color = sim_data$colors); p2
}
\references{
Roweis, Sam T. and Saul, Lawrence K., "Nonlinear Dimensionality Reduction by Locally Linear Embedding", 2000
}

