% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LTSA.R
\name{Local_TSA}
\alias{Local_TSA}
\title{Local Tangent Space Alignment}
\usage{
Local_TSA(X, k, d)
}
\arguments{
\item{X}{N x D matrix (N samples, D features).}

\item{k}{integer; Number of nearest neighbor.}

\item{d}{integer; The target dimension.}
}
\value{
N x d matrix; the projected data.
}
\description{
Local Tangent Space Alignment
}
\details{
Matlab codes were written by Zhenyue Zhang & Hongyuan Zha (2004)
and extracted from Todd Wittman's MANI: Manifold Learning Toolkit.
}
\examples{
#Simulate data
sim_data <- swiss_roll(N = 600)
library(plotly)
p1 <- plotly_3D(sim_data); p1
LTSA_data <- Local_TSA(sim_data$data, k = 8, d = 2)
p2 <- plotly_2D(LTSA_data, color = sim_data$colors); p2
}
\references{
Zhang, Z., & Zha, H. (2004). Principal Manifolds and Nonlinear Dimensionality Reduction via Tangent Space Alignment. SIAM Journal on Scientific Computing, 26(1), 313-338.

MANI: Manifold Learning Toolkit - http://macs.citadel.edu/wittman/research.html
}

