% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LTSA.R
\name{LTSA}
\alias{LTSA}
\title{Local Tangent Space Alignment (LTSA)
Written by Zhenyue Zhang & Hongyuan Zha, 2004.
Extracted from Todd Wittman's MANI: Manifold Learning Toolkit
Source: http://macs.citadel.edu/wittman/research.html
Port from Matlab to R by Jackson Kwok}
\usage{
LTSA(data, d, K, NI)
}
\description{
Local Tangent Space Alignment (LTSA)
Written by Zhenyue Zhang & Hongyuan Zha, 2004.
Extracted from Todd Wittman's MANI: Manifold Learning Toolkit
Source: http://macs.citadel.edu/wittman/research.html
Port from Matlab to R by Jackson Kwok
}

