% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maniTools.R
\docType{package}
\name{maniTools}
\alias{maniTools}
\alias{maniTools-package}
\title{Manifold learning in R}
\description{
Porting Todd Wittman's Manifold Learning Toolkit (MANI) from Matlab to R.
}
\author{
Jackson Kwok
}
\references{
http://macs.citadel.edu/wittman/research.html
}

