% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{punctured_sphere}
\alias{punctured_sphere}
\title{Punctured Sphere by Saul & Roweis}
\usage{
punctured_sphere(N, z_scale = 1)
}
\arguments{
\item{N}{integer; Number of datapoints.}

\item{z_scale}{numeric; "height" of the sphere.}
}
\description{
Punctured Sphere by Saul & Roweis
}
