% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{swiss_hole}
\alias{swiss_hole}
\title{Swiss Hole}
\usage{
swiss_hole(N, height = 1)
}
\arguments{
\item{N}{integer; Number of datapoints.}

\item{height}{numeric; height of the Swiss Roll.}
}
\description{
Swiss Hole
}
