% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{toroidal_helix}
\alias{toroidal_helix}
\title{Toroidal Helix by Coifman & Lafon}
\usage{
toroidal_helix(N, sample_rate = 1)
}
\arguments{
\item{N}{integer; Number of datapoints.}

\item{sample_rate}{numeric; Sampling rate.}
}
\description{
Toroidal Helix by Coifman & Lafon
}

