% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{twin_peaks}
\alias{twin_peaks}
\title{Twin Peaks by Saul & Roweis}
\usage{
twin_peaks(N, z_scale = 1)
}
\arguments{
\item{N}{integer; Number of datapoints.}

\item{z_scale}{numeric; "height" of the peaks.}
}
\description{
Twin Peaks by Saul & Roweis
}
