% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maniTools.R
\docType{package}
\name{maniTools}
\alias{maniTools}
\alias{maniTools-package}
\title{Manifold learning in R}
\description{
This package is about porting Todd Wittman's Matlab codes, "MANI: Manifold Learning Toolkit", to R. In particular, three algorithms are ported: Laplacian Eigenmaps, Hessian LLE (HLLE) and Local Tangent Space Alignment (LTSA). Others are collected from existing R packages, including Isomap and LLE in "RDRToolbox", diffusionMap in "diffusionMap", t-SNE in "tsne", KPCA in "kernlab" and SPE in "spe".
}
\author{
Jackson Kwok
}
\references{
http://macs.citadel.edu/wittman/research.html
}

