% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{plotly_2D}
\alias{plotly_2D}
\title{A simple 2d plot function using plotly.}
\usage{
plotly_2D(proj_data, colors)
}
\description{
A simple 2d plot function using plotly.
}
\examples{
sim_data <- swiss_roll(600)
library(plotly)
lle_data <- LLE2(sim_data$data, dim = 2, k = 8)
plotly_2D(lle_data, colors = sim_data$colors) #colors are optional
}

