% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{plotly_3D}
\alias{plotly_3D}
\title{A simple 3d plot function using plotly.}
\usage{
plotly_3D(sim_data)
}
\description{
A simple 3d plot function using plotly.
}
\examples{
sim_data <- swiss_roll(600)
library(plotly)
plotly_3D(sim_data)
}

